so many changes

+++ app/Http/Controllers/EmployeeController.php
@@ -8,7 +8,7 @@ use Illuminate\Support\Facades\DB;
 use DataTables;
 use Alert;
 use Image;
-use App\Actions\Employee\{StorePhotoAction, StoreUserAction, StoreEmployeeAction};
+use App\Actions\Employee\{StorePhotoAction, StoreUserAction, EmployeeAction};
 
 class EmployeeController extends Controller
 {
@@ -27,22 +27,23 @@ class EmployeeController extends Controller
     }
     function index()
     {
+      $employee_data = $this->employee_data->get();
       $role_data = Role::get();
-      return view('employee.index',compact('role_data'));
+      return view('employee.index',compact('role_data','employee_data'));
     }
 
-    function store(Request $request, StoreEmployeeAction $storeEmployeeAction)
+    function store(Request $request, EmployeeAction $EmployeeAction)
     {
         $employee_data = $request->validate([
             'employee_fullname' => ['required'],
             'employee_nickname' => ['required'],
-            'employee_nik' => ['required'],
+            // 'employee_nik' => ['required'],
             'employee_gender' => ['required'],
-            'employee_birthdate' => ['required'],
-            'employee_photo' => ['required'],
+            // 'employee_birthdate' => ['required'],
+            // 'employee_photo' => ['required'],
             'employee_contact' => ['required'],
             'employee_email' => ['required'],
-            'employee_address' => ['required']
+            // 'employee_address' => ['required']
         ]);
 
         $user_data = $request->validate([
@@ -52,11 +53,11 @@ class EmployeeController extends Controller
             'role_id' => ['required'],
         ]);
 
-        DB::transaction(function() use ($request, $user_data, $employee_data, $storeEmployeeAction){
-            $storeEmployeeAction->execute($request);
+        DB::transaction(function() use ($request, $user_data, $employee_data, $EmployeeAction){
+            $EmployeeAction->store($request);
         });
         Alert::success('Sukses','Data Karyawan berhasil ditambahkan !');
-        return redirect()->back()->with('success','Data Karyawan berhasil ditambahkan');
+        return redirect()->back();
     }
 
     function delete($id_employee)
@@ -65,12 +66,23 @@ class EmployeeController extends Controller
             $employee_data = $this->employee_data->find($id_employee);
             
             $user_data = User::find($employee_data->user_id);
-            unlink(public_path('/img/employee/'.$employee_data->employee_photo));
+            if($user_data->employee_image != NULL){
+                unlink(public_path('/img/employee/'.$employee_data->employee_photo));
+            }
 
             $employee_data->delete();
             $user_data->delete();
         });
         Alert::success('Sukses','Data Karyawan berhasil dihapus !');
+        return redirect()->back();
+    }
+
+    function update(Request $request, EmployeeAction $EmployeeAction, $id_employee)
+    {
+        DB::transaction(function() use ($request, $EmployeeAction, $id_employee){
+            $EmployeeAction->update($request, $id_employee);
+        });
+        Alert::success('Sukses','Data Karyawan berhasil diubah !');
         return redirect()->back();
     }
